Improved mailbox handling
This commit includes extensive changes across multiple files to
refactor the mailbox handling system and improve performance and
reliability in EtherCAT communication.

Update from Arthur Ketels 20240911.

+++ test/linux/slaveinfo/slaveinfo.c
@@ -529,7 +529,7 @@ void si_sdo(int cnt)
         printf(" CoE Object Description found, %d entries.\n",ODlist.Entries);
         for( i = 0 ; i < ODlist.Entries ; i++)
         {
-            uint8_t max_sub;
+            uint16_t max_sub;
             char name[128] = { 0 };
 
             ec_readODdescription(i, &ODlist);
